# Lazy implementqtion of bytestringsToVectors