## Performance comparison
Before:
```
benchmarking medium.csv/Foldl' over Lazy ByteString via Vector Word64
time                 1.614 s    (727.1 ms .. 2.881 s)
                     0.936 R²   (0.849 R² .. 1.000 R²)
mean                 2.034 s    (1.786 s .. 2.444 s)
std dev              357.6 ms   (0.0 s .. 379.7 ms)
variance introduced by outliers: 47% (moderately inflated)
```
After:
```
benchmarking medium.csv/Foldl' over Lazy ByteString via Vector Word64
time                 739.9 ms   (697.3 ms .. 774.5 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 723.9 ms   (712.1 ms .. 731.9 ms)
std dev              12.04 ms   (0.0 s .. 13.89 ms)
variance introduced by outliers: 19% (moderately inflated)
```